feat: updating the blubber block
