+++ blockbench/penguin/penguin.json
@@ -1,0 +1,169 @@
+{
+	"format_version": "1.8.0",
+	"animations": {
+		"animation.penguin.walk": {
+			"animation_length": 2.5833,
+			"bones": {
+				"feet R": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 0]
+						},
+						"0.5": {
+							"vector": [0, 0, 0]
+						},
+						"1.0": {
+							"vector": [0, 0, -10]
+						},
+						"1.5": {
+							"vector": [0, 0, 0]
+						},
+						"2.0": {
+							"vector": [0, 0, 0]
+						},
+						"2.5": {
+							"vector": [0, 0, -10]
+						}
+					}
+				},
+				"head": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 0]
+						},
+						"0.5": {
+							"vector": [-7.5, 0, 0]
+						},
+						"1.0": {
+							"vector": [5, 0, 0]
+						},
+						"1.5": {
+							"vector": [0, 0, 0]
+						},
+						"2.0": {
+							"vector": [-2.5, 0, 0]
+						},
+						"2.5": {
+							"vector": [5, 0, 0]
+						}
+					}
+				},
+				"body": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 0]
+						},
+						"0.5": {
+							"vector": [-5, 0, 0]
+						},
+						"1.0": {
+							"vector": [5, 0, 0]
+						},
+						"1.5": {
+							"vector": [0, 0, 0]
+						},
+						"2.0": {
+							"vector": [-5, 0, 0]
+						},
+						"2.5": {
+							"vector": [5, 0, 0]
+						}
+					}
+				},
+				"arm R": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 35]
+						},
+						"0.25": {
+							"vector": [5, 0, 45]
+						},
+						"0.5": {
+							"vector": [0, 0, 35]
+						},
+						"0.75": {
+							"vector": [5, 0, 45]
+						},
+						"1.0": {
+							"vector": [0, 0, 35]
+						},
+						"1.5": {
+							"vector": [0, 0, 35]
+						},
+						"1.75": {
+							"vector": [5, 0, 45]
+						},
+						"2.0": {
+							"vector": [0, 0, 35]
+						},
+						"2.25": {
+							"vector": [5, 0, 45]
+						},
+						"2.5": {
+							"vector": [0, 0, 35]
+						}
+					}
+				},
+				"arm L": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 32.5]
+						},
+						"0.25": {
+							"vector": [-5, 0, 42.5]
+						},
+						"0.5": {
+							"vector": [0, 0, 32.5]
+						},
+						"0.75": {
+							"vector": [-5, 0, 42.5]
+						},
+						"1.0": {
+							"vector": [0, 0, 32.5]
+						},
+						"1.5": {
+							"vector": [0, 0, 32.5]
+						},
+						"1.75": {
+							"vector": [-5, 0, 42.5]
+						},
+						"2.0": {
+							"vector": [0, 0, 32.5]
+						},
+						"2.25": {
+							"vector": [-5, 0, 42.5]
+						},
+						"2.5": {
+							"vector": [0, 0, 32.5]
+						}
+					}
+				},
+				"feet L": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 0]
+						},
+						"0.5": {
+							"vector": [0, 0, -7.5]
+						},
+						"1.0": {
+							"vector": [0, 0, 0]
+						},
+						"1.5": {
+							"vector": [0, 0, 0]
+						},
+						"2.0": {
+							"vector": [0, 0, -7.5]
+						},
+						"2.5": {
+							"vector": [0, 0, 0]
+						}
+					}
+				}
+			}
+		},
+		"animation.penguin.idle": {
+			"loop": true
+		}
+	}
+}
