+++ src/main/java/net/neetcoders/miraculousmanatee/block/BlubberBlock.java
@@ -1,4 +1,4 @@
-package net.neetcoders.miraculousmanatee;
+package net.neetcoders.miraculousmanatee.block;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
