+++ src/main/resources/assets/miraculousmanatee/blockstates/blubber_block.json
@@ -1,5 +1,5 @@
 {
   "variants": {
-    "": { "model": "miraculousmanatee:block/blubber_block_world" }
+    "": { "model": "miraculousmanatee:block/blubber_block" }
   }
 }
