+++ src/main/resources/assets/miraculousmanatee/models/block/blubber_block.json
@@ -1,6 +1,23 @@
 {
-  "parent": "minecraft:block/cube_all",
+  "credit": "Made with Blockbench",
+  "texture_size": [64, 64],
   "textures": {
-    "all": "miraculousmanatee:block/blubber_block"
-  }
+    "0": "miraculousmanatee:block/blubber_block",
+    "particle": "miraculousmanatee:block/blubber_block"
+  },
+  "elements": [
+    {
+      "name": "blubber",
+      "from": [0, 0, 0],
+      "to": [16, 16, 16],
+      "faces": {
+        "north": {"uv": [4, 4, 8, 8], "texture": "#0"},
+        "east": {"uv": [0, 4, 4, 8], "texture": "#0"},
+        "south": {"uv": [12, 4, 16, 8], "texture": "#0"},
+        "west": {"uv": [8, 4, 12, 8], "texture": "#0"},
+        "up": {"uv": [8, 4, 4, 0], "texture": "#0"},
+        "down": {"uv": [12, 0, 8, 4], "texture": "#0"}
+      }
+    }
+  ]
 }
